Split ProposalService and BlockchainProposalService into smaller parts
Split ProposalService into the following interfaces:
- ProposalReader
- ProposalWriter
- UserActivityTracker (might be good to separate from ProposalService)

BlockchainProposalService is now split into these components:
- ProposalReader
- ProposalWriter
- UserActivityTracker
- ProposalParser
- EventProvider
