+++ __tests__/new-proposal-dialog.test.tsx
@@ -10,8 +10,7 @@ import {
 } from "@testing-library/react";
 import { useProposals } from "../hooks/use-proposals";
 import { ProposalsContextValue } from "@/context/proposals-context";
-import { InMemoryProposalFileService } from "@/lib/file-upload";
-import { InMemoryProposalService } from "@/lib/proposal-services/in-memory-proposal-service";
+import { InMemoryProposalService } from "@/lib/proposal-services/in-memory/in-memory-proposal-service";
 import { User } from "@/types/proposal";
 import { NewProposalDialog } from "@/components/new-proposal-dialog";
 import userEvent from "@testing-library/user-event";
