+++ lib/proposal-services/blockchain/blockchain-proposal-parser.ts
@@ -1,0 +1,173 @@
+import {
+  AddVoterVotableItem,
+  IsUIAddVoterVotableItem,
+  Proposal,
+  UIAddVoterVotableItem,
+  UIProposal,
+  UIVotableItem,
+  VotableItem,
+  VoteEvent,
+} from "../../../types/proposal";
+import {
+  convertAddressToName,
+  convertGovernorState,
+  getNewVoterProposalDescription,
+} from "../../utils";
+import { ProposalFileService } from "../../file-upload";
+import { EvsdGovernor } from "@/typechain-types";
+import { ProposalCreatedEvent } from "@/typechain-types/contracts/EvsdGovernor";
+
+interface ChainData {
+  title: string;
+  description: string;
+  type: "proposal" | "voteItem";
+}
+
+export interface ProposalChainData extends ChainData {
+  type: "proposal";
+  fileHash: string;
+}
+export interface VotableItemChainData extends ChainData {
+  type: "voteItem";
+  parentProposalId: string;
+  index: number;
+}
+export interface AddVoterVotableItemChainData extends VotableItemChainData {
+  newVoterAddress: string;
+}
+
+export const isProposalChainData = (
+  data: ChainData
+): data is ProposalChainData => data.type === "proposal";
+export const isVotableItemChainData = (
+  data: ChainData
+): data is VotableItemChainData => data.type === "voteItem";
+export const isAddVoterVotableItemChainData = (
+  data: ChainData
+): data is AddVoterVotableItemChainData =>
+  isVotableItemChainData(data) && "newVoterAddress" in data;
+
+export class BlockchainProposalParser {
+  private readonly governor: EvsdGovernor;
+  private readonly fileService: ProposalFileService;
+
+  constructor(governor: EvsdGovernor, fileService: ProposalFileService) {
+    this.governor = governor;
+    this.fileService = fileService;
+  }
+
+  public deserializeChainData(data: string): ChainData {
+    return JSON.parse(data) as ChainData;
+  }
+
+  public async parseProposal(
+    proposalData: ProposalChainData,
+    args: ProposalCreatedEvent.OutputTuple & ProposalCreatedEvent.OutputObject
+  ): Promise<Proposal> {
+    const proposalId = args.proposalId;
+    const voteStart = new Date(Number(args.voteStart) * 1000);
+    const stateIndex = Number(
+      (await this.governor.state(proposalId)) as bigint
+    );
+    const proposalState = convertGovernorState(stateIndex);
+    const deadline = await this.governor.proposalDeadline(proposalId);
+    const closesAt = new Date(Number(deadline) * 1000);
+    // Create a proposal with an empty itemsToVote array - it will be filled after all of the VotableItems arrive
+    return {
+      id: proposalId,
+      title: proposalData.title,
+      description: proposalData.description,
+      author: {
+        address: args.proposer,
+        name: convertAddressToName(args.proposer),
+      },
+      file:
+        proposalData.fileHash !== ""
+          ? await this.fileService.fetch(proposalData.fileHash)
+          : undefined,
+      dateAdded: voteStart,
+      status: proposalState,
+      closesAt: closesAt,
+      voteItems: [],
+    };
+  }
+  public async parseVotableItem(
+    deserializedData: VotableItemChainData,
+    args: ProposalCreatedEvent.OutputTuple & ProposalCreatedEvent.OutputObject,
+    voteEventsForId: Record<string, VoteEvent[]>
+  ): Promise<VotableItem> {
+    const proposalId = args.proposalId;
+    const proposalIdStr = proposalId.toString();
+
+    // Convert an array of VoteEvents into a map of voter -> voteEvent
+    const votesForAddress =
+      proposalIdStr in voteEventsForId
+        ? voteEventsForId[proposalIdStr].reduce<Map<string, VoteEvent>>(
+            (acc, item) => {
+              acc.set(item.voter.address, item);
+              return acc;
+            },
+            new Map()
+          )
+        : new Map<string, VoteEvent>();
+
+    return {
+      id: proposalId,
+      title: deserializedData.title,
+      description: deserializedData.description,
+      userVotes: votesForAddress,
+    };
+  }
+  public async parseAddVoterVotableItem(
+    deserializedData: AddVoterVotableItemChainData,
+    args: ProposalCreatedEvent.OutputTuple & ProposalCreatedEvent.OutputObject,
+    voteEventsForId: Record<string, VoteEvent[]>
+  ): Promise<AddVoterVotableItem> {
+    const votableItem = await this.parseVotableItem(
+      deserializedData,
+      args,
+      voteEventsForId
+    );
+    return {
+      ...votableItem,
+      newVoterAddress: deserializedData.newVoterAddress,
+    };
+  }
+  public async serializeProposal(proposal: UIProposal | Proposal) {
+    if (proposal.file) {
+      throw new Error("Proposal file upload not supported yet");
+    }
+
+    const serializationData: ProposalChainData = {
+      type: "proposal",
+      title: proposal.title,
+      description: proposal.description,
+      fileHash: "",
+    };
+    return JSON.stringify(serializationData);
+  }
+  public serializeVotableItem(
+    item: UIVotableItem | UIAddVoterVotableItem,
+    parentId: bigint,
+    index: number
+  ): string {
+    const serializationData:
+      | VotableItemChainData
+      | AddVoterVotableItemChainData = IsUIAddVoterVotableItem(item)
+      ? {
+          type: "voteItem",
+          parentProposalId: parentId.toString(),
+          newVoterAddress: item.newVoterAddress,
+          index: 0,
+          ...getNewVoterProposalDescription(item.newVoterAddress),
+        }
+      : {
+          type: "voteItem",
+          title: item.title,
+          description: item.description,
+          parentProposalId: parentId.toString(),
+          index,
+        };
+    return JSON.stringify(serializationData);
+  }
+}
