+++ lib/proposal-services/proposal-service.ts
@@ -4,21 +4,33 @@ import {
   VotableItem,
   VoteOption,
 } from "@/types/proposal";
-import { onProposalsChangedUnsubscribe } from "./blockchain-proposal-service";
 import {
   UserActivityEventVote,
   UserActivityEventProposal,
 } from "@/components/user-activity/user-activity";
 
-export interface ProposalService {
+export type onProposalsChangedUnsubscribe = () => void;
+
+export interface ProposalReader {
   getProposals: () => Promise<Proposal[]>;
-  uploadProposal: (proposal: UIProposal) => Promise<bigint>;
-  voteForItem: (item: VotableItem, vote: VoteOption) => Promise<void>;
-  cancelProposal(proposal: Proposal): Promise<boolean>;
   onProposalsChanged(
     callback: (newProposals: Proposal[]) => void
   ): onProposalsChangedUnsubscribe;
-  getAllUserActivity(): Promise<
+}
+
+export interface UserActivityTracker {
+  getAllUserActivity: () => Promise<
     (UserActivityEventVote | UserActivityEventProposal)[]
   >;
 }
+
+export interface ProposalWriter {
+  uploadProposal: (proposal: UIProposal) => Promise<bigint>;
+  voteForItem: (item: VotableItem, vote: VoteOption) => Promise<void>;
+  cancelProposal(proposal: Proposal): Promise<boolean>;
+}
+
+export interface ProposalService
+  extends ProposalReader,
+    ProposalWriter,
+    UserActivityTracker {}
