+++ lib/utils.ts
@@ -14,6 +14,8 @@ import {
 } from "../types/proposal";
 import { addressNameMap } from "../constants/address-name-map";
 import { STRINGS } from "../constants/strings";
+import { ethers } from "ethers";
+import { EvsdToken } from "@/typechain-types";
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
@@ -223,8 +225,12 @@ async function areFilesEqual(file1?: File, file2?: File): Promise<boolean> {
   }
   if (file1 && file2) {
     // Try some early outs
-    if (file1.size !== file2.size) return false;
-    if (file1.name === file2.name && file1 === file2) return true;
+    if (file1.size !== file2.size) {
+      return false;
+    }
+    if (file1.name === file2.name && file1 === file2) {
+      return true;
+    }
 
     // Read and compare buffers byte by byte
     const [buffer1, buffer2] = await Promise.all([
@@ -234,7 +240,9 @@ async function areFilesEqual(file1?: File, file2?: File): Promise<boolean> {
     const view1 = new Uint8Array(buffer1);
     const view2 = new Uint8Array(buffer2);
     for (let i = 0; i < view1.length; i++) {
-      if (view1[i] !== view2[i]) return false;
+      if (view1[i] !== view2[i]) {
+        return false;
+      }
     }
 
     return true;
@@ -251,4 +259,17 @@ export async function areProposalsEqual(
     proposal.description === uiProposal.description &&
     (await areFilesEqual(proposal.file, uiProposal.file))
   );
+}
+
+export async function getTransferTokenCalldata(
+  token: EvsdToken,
+  newVoterAddress: string
+) {
+  const decimals = await token.decimals();
+  const oneToken = ethers.parseUnits("1", decimals);
+  const transferCalldata = token.interface.encodeFunctionData("transfer", [
+    newVoterAddress,
+    oneToken,
+  ]);
+  return transferCalldata;
 }
