+++ test/utils.ts
@@ -18,7 +18,7 @@ import {
   VoteOption,
 } from "../types/proposal";
 import seedrandom from "seedrandom";
-import { BlockchainProposalService } from "../lib/proposal-services/blockchain-proposal-service";
+import { BlockchainProposalService } from "../lib/proposal-services/blockchain/blockchain-proposal-service";
 import { InMemoryProposalFileService } from "../lib/file-upload";
 import { ProposalService } from "@/lib/proposal-services/proposal-service";
 export const rng = seedrandom("42");
@@ -94,8 +94,6 @@ export async function deployAndCreateMocks(): Promise<TestInitData> {
   const { token, governor } = await deployContracts(
     owner as unknown as ethers.Signer
   );
-  const evsdGovernor = governor;
-  const evsdToken = token;
   await distributeVotingRights(
     owner as unknown as ethers.Signer,
     token,
@@ -112,21 +110,21 @@ export async function deployAndCreateMocks(): Promise<TestInitData> {
   const registeredVoterProposalServices = voters.map(
     (voter) =>
       new BlockchainProposalService(
+        governor,
+        token,
         voter as unknown as ethers.Signer,
         fileService,
-        hardhat.ethers.provider,
-        governorAddress,
-        tokenAddress
+        hardhat.ethers.provider
       )
   );
   const unregisteredVoter = await getUnregisteredVoter();
   const unregisteredVoterProposalServices = [
     new BlockchainProposalService(
+      governor,
+      token,
       unregisteredVoter as unknown as ethers.Signer,
       fileService,
-      hardhat.ethers.provider,
-      governorAddress,
-      tokenAddress
+      hardhat.ethers.provider
     ),
   ];
   const addVoterVoteItem = { newVoterAddress: unregisteredVoter.address };
@@ -136,8 +134,8 @@ export async function deployAndCreateMocks(): Promise<TestInitData> {
     unregisteredVoterProposalServices,
     addVoterVoteItem,
     votingPeriod,
-    evsdGovernor,
-    evsdToken,
+    evsdGovernor: governor,
+    evsdToken: token,
     unregisteredVoterAddress: unregisteredVoter.address,
   };
   return initData;
